+++ plugin-lifecycle-lint-rules/src/main/java/com/mapbox/maps/lint/lifecycle/LifecycleMethodDetector.kt
@@ -23,13 +23,14 @@ class LifecycleMethodDetector : Detector(), Detector.UastScanner {
   companion object Issues {
     private const val ON_START = "onStart"
     private const val ON_STOP = "onStop"
+    private const val ON_PAUSE = "onPause"
     private const val ON_DESTROY = "onDestroy"
     private const val ON_LOW_MEMORY = "onLowMemory"
     private const val ISSUE_ID = "Lifecycle"
     private const val BRIEF_DESCRIPTION = "Redundant method call"
     private const val PRIORITY = 8
     private const val REPORT_MESSAGE =
-      "It's no longer required to manually invoke onStart/onStop/onDestroy/onLowMemory MapView lifecycle methods explicitly. The Mapbox Maps SDK for Android depends on the plugin-lifecycle that leverages androidx.lifecycle to make the MapView a lifecycle aware component. This behavior is automatically enforced when appcompact 1.3.0+ is used. The plugin-lifecycle can be removed from class path to opt-out of the associated behavior."
+      "It's no longer required to manually invoke onStart/onStop/onPause/onDestroy/onLowMemory MapView lifecycle methods explicitly. The Mapbox Maps SDK for Android depends on the plugin-lifecycle that leverages androidx.lifecycle to make the MapView a lifecycle aware component. This behavior is automatically enforced when appcompact 1.3.0+ is used. The plugin-lifecycle can be removed from class path to opt-out of the associated behavior."
 
     private val IMPLEMENTATION = Implementation(
       LifecycleMethodDetector::class.java, Scope.JAVA_FILE_SCOPE