+++ sdk-base/src/main/java/com/mapbox/maps/MapboxLifecycleObserver.kt
@@ -23,4 +23,9 @@ interface MapboxLifecycleObserver {
    * Called to reduce memory use
    */
   fun onLowMemory()
+
+  /**
+   * Called when the activity is resumed
+   */
+  fun onResume()
 }